+++ front-end-project-for-backend-week/lambdaNotes/serverDerp/server.js
@@ -2,12 +2,128 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const morgan = require('morgan');
 
-const Note = require('./model');
+const session = require('express-session');
+const bcrypt = require('bcrypt');
+const cors = require('cors');
+const mongoose = require('mongoose');
 
+const corsOptions = {
+  origin: 'http://localhost:2323',
+  credentials: true,
+};
 
 const server = express();
+server.use(cors(corsOptions));
 server.use(bodyParser.json());
-server.use(morgan('combined'));
+server.use(
+  session({
+    secret: 'aNtIdestablishmentarianism',
+    resave: true,
+    saveUninitialized: true,
+  }),
+);
+
+mongoose 
+  .connect('mongodb://localhost/backEndProj')
+  .then(() => console.log('successfully connected to database'))
+  .catch(() => console.log('Connection failed'));
+
+const Note = require('./model');
+const User = require('./userModel')
+
+//Middleware
+
+const sendUserError = (err, res) => {
+  res.status(STATUS_USER_ERROR);
+  if (err && err.message) {
+    res.json({ message: err.message, stack: err.stack });
+  } else {
+    res.json({ error: err });
+  }
+};
+
+const loggedIn = (req, res, next) => {
+  const { username } = req.session;
+  console.log(req.session);
+  if (!username) {
+    sendUserError('User is not logged in', res);
+    return;
+  }
+  User.findOne({ username }, (err, user) => {
+    if (err) {
+      sendUserError(err, res);
+    } else if (!user) {
+      sendUserError('User does exist', res);
+    } else {
+      req.user = user;
+      next();
+    }
+  });
+};
+
+
+const authenticate = (req, res, next) => {
+  const path = req.path;
+  if (/restricted/.test(path)) {
+
+    if (!req.session.username) {
+      sendUserError('user not authorized', res);
+      return;
+    }
+  }
+  next();
+};
+server.use(authenticate);
+
+
+
+
+//Auth Routes
+
+server.post('/login', (req, res) => {
+  const { username, password } = req.body;
+  if (!username) {
+    sendUserError('username undefined', res);
+    return;
+  }
+  User.findOne({ username }, (err, user) => {
+    if (err || user === null) {
+      sendUserError('No user found at that id', res);
+      return;
+    }
+    const hashedPw = user.password;
+    bcrypt
+      .compare(password, hashedPw)
+      .then((response) => {
+        if (!response) throw new Error();
+        req.session.username = username;
+        req.user = user;
+      })
+      .then(() => {
+        res.json({ success: true, user});
+      })
+      .catch((error) => {
+        return sendUserError('User does not exist at that id ', res);
+      });
+  });
+});
+
+server.post('/register', (req, res) => {
+  User.create(req.body)
+  .then(user => res.status(200).json(user))
+  .catch(err => res.json({ msg:'Could not create User', err}))
+});
+
+server.post('/logout', (req, res) => {
+  if (!req.session.username) {
+    sendUserError('User is not logged in', res);
+    return;
+  }
+  req.session.username = null;
+  res.json({ success: true });
+});
+
+//Note Routes
 
 server.post('/note/create', (req, res) => {
   const { title, content } = req.body;
